EWCE parallel

Runs EWCE in parallel on multiple gene lists.

ewce_para(
  ctd,
  gene_data,
  list_name_column = "Phenotype",
  gene_column = "Gene",
  list_names = unique(gene_data[[list_name_column]]),
  bg = unique(gene_data[[gene_column]]),
  reps = 100,
  annotLevel = 1,
  genelistSpecies = "human",
  sctSpecies = "human",
  save_dir_tmp = tempdir(),
  seed = 2022,
  cores = 1,
  verbose = FALSE
)

Arguments

ctd

Cell Type Data List generated using generate_celltype_data.

gene_data

data frame of gene list names and genes (see get_gene_list).

list_name_column

The name of the gene_data column that has the gene list names.

gene_column

The name of the gene_data column that contains the genes.

list_names

character vector of gene list names.

bg

List of gene symbols containing the background gene list (including hit genes). If bg=NULL, an appropriate gene background will be created automatically.

reps

Number of random gene lists to generate (Default: 100, but should be >=10,000 for publication-quality results).

annotLevel

An integer indicating which level of sct_data to analyse (Default: 1).

genelistSpecies

Species that hits genes came from (no longer limited to just "mouse" and "human"). See list_species for all available species.

sctSpecies

Species that sct_data came from (no longer limited to just "mouse" and "human"). See list_species for all available species.

save_dir_tmp

Folder to save intermediate results files to (one file per gene list). Set to NULL to skip saving temporary files.

seed

Used to set the seed for replicable results.

cores

The number of cores to run in parallel (e.g. 8) int.

verbose

Print messages.

Value

Paths to saved results at "(save_dir)/(list_name).rds" (when save_dir!=NULL), or a nested list of results (when save_dir==NULL).

Examples

gene_data <- HPOExplorer::load_phenotype_to_genes()
#> Importing existing file: /github/home/.cache/R/HPOExplorer/data/phenotype_to_genes.txt
ctd <- MultiEWCE::load_example_ctd()
#> ℹ All local files already up-to-date!
list_names <- unique(gene_data$Phenotype)[seq_len(3)]
res_files <- ewce_para(ctd = ctd,
                       gene_data = gene_data,
                       list_names = list_names,
                       reps = 10)
#> Converting to sparse matrix.
#> Converting to sparse matrix.
#> Converting to sparse matrix.