Generate results

Generates EWCE results on multiple gene lists in parallel by calling ewce_para. It allows you to stop the analysis and then continue later from where you left off as it checks the results output directory for finished gene lists and removes them from the input. It also excludes gene lists with less than 4 unique genes (which cause errors in EWCE analysis).

gen_results(
  ctd,
  gene_data,
  list_name_column = "Phenotype",
  gene_column = "Gene",
  list_names = unique(gene_data[[list_name_column]]),
  bg = unique(gene_data[[gene_column]]),
  force_new = FALSE,
  reps = 100,
  annotLevel = 1,
  genelistSpecies = "human",
  sctSpecies = "human",
  cores = 1,
  save_dir_tmp = NULL,
  save_dir = tempdir(),
  verbose = 1
)

Arguments

ctd

Cell Type Data List generated using generate_celltype_data.

gene_data

data frame of gene list names and genes (see get_gene_lists).

list_name_column

The name of the gene_data column that has the gene list names.

gene_column

The name of the gene_data column that contains the genes.

list_names

character vector of gene list names.

bg

List of gene symbols containing the background gene list (including hit genes). If bg=NULL, an appropriate gene background will be created automatically.

force_new

Overwrite previous results in the save_dir_tmp.

reps

Number of random gene lists to generate (Default: 100, but should be >=10,000 for publication-quality results).

annotLevel

An integer indicating which level of sct_data to analyse (Default: 1).

genelistSpecies

Species that hits genes came from (no longer limited to just "mouse" and "human"). See list_species for all available species.

sctSpecies

Species that sct_data came from (no longer limited to just "mouse" and "human"). See list_species for all available species.

cores

The number of cores to run in parallel (e.g. 8) int.

save_dir_tmp

Folder to save intermediate results files to (one file per gene list). Set to NULL to skip saving temporary files.

save_dir

Folder to save merged results in.

verbose

Print messages.

Value

All results as a dataframe.

Details

The gene_data should be a data frame that contains a column of gene list names (e.g. the column may be called "Phenotype"), and a column of genes (e.g. "Gene"). For example:

Phenotype	Gene

"Abnormal heart"	gene X

"Abnormal heart"	gene Y

"Poor vision"	gene Z

"Poor vision"	gene Y

etc...

For more information on this see docs for get_gene_list (get_gene_lists).

Examples

gene_data <- HPOExplorer::load_phenotype_to_genes()
#> Importing existing file: ... phenotype_to_genes.txt
list_names <- unique(gene_data$Phenotype)[seq_len(5)]
ctd <- load_example_ctd()
all_results <- gen_results(ctd = ctd,
                           gene_data = gene_data,
                           list_names = list_names,
                           reps = 10)
#> Validating gene lists..
#> 4 / 5 gene lists are valid.
#> Converting to sparse matrix.
#> Converting to sparse matrix.
#> Converting to sparse matrix.
#> Converting to sparse matrix.
#> 
#> Saving results ==> /tmp/RtmpKoejJR/results_2023-02-09_13-08-07.080065.rds