get_gene_list

This returns a character vector of genes (a gene list). The gene_data should be a data frame that contains a column of gene list names (e.g. the column may be called "Phenotype"), and a column of genes (e.g. "Gene"). For example, lets call the following data frame phenotype_to_genes:

get_gene_list(
  list_name,
  gene_data,
  list_name_column = "Phenotype",
  gene_column = "Gene"
)

Arguments

list_name
The name of the gene list of interest string

gene_data
The data frame of gene list names and associated genes (data.frame)

list_name_column
The name of the column in gene_data that contains gene list names string

gene_column
The name of the column in gene_data that contains the genes string

Value

A charcter vector of genes associated with the selected list_name

Details

Phenotype	Gene

"Abnormal heart"	gene X

"Abnormal heart"	gene Y

"Poor vision"	gene Z

"Poor vision"	gene Y

etc...

In the example above, if we wanted to extract genes related to "Abnormal heart", we would run the following example:

Examples


phenotype_to_genes <- HPOExplorer::load_phenotype_to_genes("phenotype_to_genes.txt")

height_genes <- get_gene_list(list_name = "Abnormality of body height",
                             gene_data = phenotype_to_genes,
                             list_name_column = "Phenotype",
                             gene_column = "Gene")