Get unfinished list names

Gets the unfinished gene list names. It reads the file names in the results directory and uses this to deduce which gene lists have already been analysed. This means you can pause the analysis of multiple gene lists and it will not re-analyse the already completed ones when you start again.

get_unfinished_list_names(list_names, save_dir_tmp)

Arguments

list_names

character vector of gene list names.

save_dir_tmp

Folder to save intermediate results files to (one file per gene list). Set to NULL to skip saving temporary files.

Value

A character vector of list_names that still need to be analysed.

Examples

gene_data <- HPOExplorer::load_phenotype_to_genes()
#> Importing existing file: /github/home/.cache/R/HPOExplorer/data/phenotype_to_genes.txt
list_names <- unique(gene_data$Phenotype)[seq_len(3)]
save_dir_tmp <- file.path(tempdir(),"results")
res_files <- ewce_para(ctd = ctd,
                       gene_data = gene_data,
                       list_names = list_names,
                       reps = 10,
                       save_dir_tmp = save_dir_tmp)
#> Error in check_bootstrap_args(sct_data = sct_data, hits = hits, annotLevel = annotLevel,     reps = reps, controlledCT = controlledCT): Must provide valid CellTypeDataset to sct_data.
unfinished <- get_unfinished_list_names(list_names = list_names,
                                        save_dir_tmp = save_dir_tmp)