Create interactive network plot start to finish

This puts all the functions together from getting the subset of results to creating the final interactive plot.

ggnetwork_plot_full(
  cell_type,
  results = load_example_results(),
  hpo = HPOExplorer::get_hpo(),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  q_threshold = 5e-04,
  fold_threshold = 1,
  columns = list(Phenotype = "Phenotype", ID = "HPO_ID", ontLvl_genes =
    "ontLvl_geneCount_ratio", Description = "description", CellType = "CellType", p =
    "p", q = "q", fold_change = "fold_change"),
  colour_var = "fold_change",
  interactive = TRUE,
  verbose = TRUE
)

Arguments

cell_type

The cell type of interest to be plotted.

results

The cell type-phenotype enrichment results generated by gen_results and merged together with merge_results.

hpo

Human Phenotype Ontology object, loaded from ontologyIndex.

phenotype_to_genes

Output of load_phenotype_to_genes mapping phenotypes to gene annotations.

q_threshold

The q value threshold to subset the results by.

fold_threshold

The minimum fold change in specific expression to subset the results by.

columns

A named vector of columns in phenos to add to the hoverdata via make_hoverboxes.

colour_var

column name of the variable to be mapped to colour \<string\>

interactive

Make the plot interactive with ggplotly.

verbose

Print messages.

Value

A named list of outputs, including a interactive network plot of the selected subset of results from RD EWCE analysis.

Examples

res_list <- ggnetwork_plot_full(cell_type = "Microglia")
#> ggnetwork_plot_full
#> Subsetting results by phenotype.
#> get_cell_ontology
#> Importing existing file: /github/home/.cache/R/HPOExplorer/data/phenotype_to_genes.txt
#> 9 associations remain after filtering.
#> Creating adjacency matrix for 9 terms.
#> Making hoverboxes from: 'Phenotype', 'HPO_ID', 'description', 'CellType', 'p', 'q', 'fold_change'
#> Making phenotype network object.
#> Getting ontology level for 9 HPO IDs.
#> Creating ggnetwork plot.