Prioritise target genes

Prioritise target genes based on a procedure:

Include only results at q<=0.05.

Include only results at fold_change>=1.

Include only terminally differentiated cells.

priortise_targets(
  results,
  ctd,
  annotLevel = 1,
  q_threshold = 0.05,
  fold_threshold = 1,
  keep_celltypes = NULL,
  keep_tiers = c(1, 2),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  hpo = HPOExplorer::get_hpo()
)

Arguments

results

The cell type-phenotype enrichment results generated by gen_results and merged together with merge_results.

ctd

Cell Type Data List generated using generate_celltype_data.

annotLevel

An integer indicating which level of sct_data to analyse (Default: 1).

q_threshold

The q value threshold to subset the results by.

fold_threshold

The minimum fold change in specific expression to subset the results by.

keep_celltypes

Cell type to keep.

keep_tiers

Tiers from hpo_tiers to keep.

phenotype_to_genes

Output of load_phenotype_to_genes mapping phenotypes to gene annotations.

hpo

Human Phenotype Ontology object, loaded from ontologyIndex.

Examples

results <- load_example_results()
ctd <- load_example_ctd()