EWCE parallel

Runs EWCE in parallel on multiple gene lists.

ewce_para(
  list_names,
  gene_data,
  list_name_column = "Phenotype",
  gene_column = "Gene",
  results_directory,
  ctd_file,
  background_genes,
  bootstrap_reps,
  annotation_Level,
  genes_Species,
  ctd_Species,
  cores
)

Arguments

list_names
character vector of gene list names

gene_data
data frame of gene list names and genes (see ?get_gene_list)

list_name_column
The name of the gene_data column that has the gene list names

gene_column
The name of the gene_data column that contains the genes

results_directory
The desired output filepath for results to be saved

ctd_file
The cell type data object for EWCE analysis (see EWCE docs)

background_genes
The background geneset for EWCE analysis (see EWCE docs)

bootstrap_reps
The number of bootstrap reps int (e.g. 100000)

annotation_Level
The level of desired cell resolution from the CTD

genes_Species
The species of gene lists string "human" or "mouse"

ctd_Species
"human" or "mouse" string

cores
The number of cores to run in parallel (e.g. 8) int

Value

True if analysis was sucessful. The results will then be saved at "(results_directory)/(list_name).rds"

Examples

gene_data <- HPOExplorer::load_phenotype_to_genes("phenotype_to_genes.txt")
ctd <- MultiEWCE::load_example_CTD()

#> Error: 'load_example_CTD' is not an exported object from 'namespace:MultiEWCE'

list_names <- unique(gene_data$Phenotype)[1:10]
results <- ewce_para(list_names, gene_data, results_directory ="results",ctd_file = ctd,
          background_genes = unique(gene_data$Gene), bootstrap_reps = 10,
          annotation_Level = 1, genes_Species = "human", ctd_Species = "human",
          cores = 1)

#> [1] "Abnormality of body height"
#> Error in unique(is.na(sct_data)) : object 'ctd' not found
#> [1] "Multicystic kidney dysplasia"

#> Warning: restarting interrupted promise evaluation

#> Error in unique(is.na(sct_data)) : object 'ctd' not found
#> [1] "Autosomal dominant inheritance"

#> Warning: restarting interrupted promise evaluation

#> Error in unique(is.na(sct_data)) : object 'ctd' not found
#> [1] "Autosomal recessive inheritance"

#> Warning: restarting interrupted promise evaluation

#> Error in unique(is.na(sct_data)) : object 'ctd' not found
#> [1] "Abnormal morphology of female internal genitalia"

#> Warning: restarting interrupted promise evaluation

#> Error in unique(is.na(sct_data)) : object 'ctd' not found
#> [1] "Functional abnormality of the bladder"

#> Warning: restarting interrupted promise evaluation

#> Error in unique(is.na(sct_data)) : object 'ctd' not found
#> [1] "Recurrent urinary tract infections"

#> Warning: restarting interrupted promise evaluation

#> Error in unique(is.na(sct_data)) : object 'ctd' not found
#> [1] "Neurogenic bladder"

#> Warning: restarting interrupted promise evaluation

#> Error in unique(is.na(sct_data)) : object 'ctd' not found
#> [1] "Urinary urgency"

#> Warning: restarting interrupted promise evaluation

#> Error in unique(is.na(sct_data)) : object 'ctd' not found
#> [1] "Hypoplasia of the uterus"

#> Warning: restarting interrupted promise evaluation

#> Error in unique(is.na(sct_data)) : object 'ctd' not found