EWCE parallel

Runs EWCE in parallel on multiple gene lists.

ewce_para(
  list_names,
  gene_data,
  list_name_column = "Phenotype",
  gene_column = "Gene",
  results_directory,
  ctd_file,
  background_genes,
  bootstrap_reps,
  annotation_Level,
  genes_Species,
  ctd_Species,
  cores
)

Arguments

list_names
character vector of gene list names

gene_data
data frame of gene list names and genes (see ?get_gene_list)

list_name_column
The name of the gene_data column that has the gene list names

gene_column
The name of the gene_data column that contains the genes

results_directory
The desired output filepath for results to be saved

ctd_file
The cell type data object for EWCE analysis (see EWCE docs)

background_genes
The background geneset for EWCE analysis (see EWCE docs)

bootstrap_reps
The number of bootstrap reps <int> (e.g. 100000)

annotation_Level
The level of desired cell resolution from the CTD

genes_Species
The species of gene lists <string> "human" or "mouse"

ctd_Species
"human" or "mouse" <string>

cores
The number of cores to run in parallel (e.g. 8) <int>

Value

True if analysis was sucessful, saves results at "<results_directory>/<list_name>.rds"