Prioritise target genes

Prioritise target genes based on a procedure:

Association-level: q_threshold: Keep only results at q<=0.05.

Association-level: fold_threshold: Keep only results at fold_change>=1.

Phenotype-level: keep_ont_levels: Keep only phenotypes at certain absolute ontology levels within the HPO.

Phenotype-level: keep_onsets: Keep only phenotypes with certain age of onsets.

Phenotype-level: keep_tiers: Keep only phenotypes with high severity Tiers.

Phenotype-level: severity_threshold: Keep only phenotypes with mean Severity equal to or below the threshold.

Phenotype-level: pheno_frequency_threshold: Keep only phenotypes with mean frequency equal to or above the threshold (i.e. how frequently a phenotype is associated with any diseases in which it occurs).

Cell type-level: keep_celltypes: Keep only terminally differentiated cell types.

Gene-level: keep_seqnames: Remove genes on non-standard chromosomes.

Gene-level: gene_size: Keep only genes <4.3kb in length.

Gene-level: keep_biotypes: Keep only genes belonging to certain biotypes.

Gene-level: gene_frequency_threshold: Keep only genes at or above a certain mean frequency threshold (i.e. how frequently a gene is associated with a given phenotype when observed within a disease).

Gene-level: keep_specificity_quantiles: Keep only genes in top specificity quantiles from the cell type dataset (ctd).

Gene-level: keep_mean_exp_quantiles: Keep only genes in top mean expression quantiles from the cell type dataset (ctd).

All levels: top_n: Sort candidate targets by a preferred order of metrics and only return the top N targets per cell type-phenotype combination.

prioritise_targets(
  results = load_example_results(),
  ctd = load_example_ctd(),
  annotLevel = 1,
  q_threshold = 0.05,
  fold_threshold = 1,
  keep_ont_levels = NULL,
  keep_onsets = HPOExplorer::list_onsets(exclude_onsets = c("Antenatal", "Fetal")),
  keep_tiers = c(1, 2),
  severity_threshold = c(2, NA),
  pheno_frequency_threshold = 25,
  keep_celltypes = terminal_celltypes()$CellType,
  keep_seqnames = c(seq_len(22), "X", "Y"),
  gene_size = list(min = 0, max = 4300),
  gene_frequency_threshold = c(10, NA),
  keep_biotypes = NULL,
  keep_specificity_quantiles = seq(39, 40),
  keep_mean_exp_quantiles = keep_specificity_quantiles,
  sort_cols = c(tier = 1, tier_auto = 1, Severity_score_mean = 1, q = 1, fold_change =
    -1, specificity_quantile = -1, mean_exp_quantile = -1, specificity = -1, mean_exp =
    -1, pheno_freq_mean = -1, gene_freq_mean = -1, width = 1),
  top_n = 20,
  group_vars = c("HPO_ID", "CellType"),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  hpo = HPOExplorer::get_hpo(),
  return_report = TRUE,
  verbose = TRUE
)

Arguments

results

The cell type-phenotype enrichment results generated by gen_results and merged together with merge_results.

ctd

Cell Type Data List generated using generate_celltype_data.

annotLevel

An integer indicating which level of sct_data to analyse (Default: 1).

q_threshold

The q value threshold to subset the results by.

fold_threshold

The minimum fold change in specific expression to subset the results by.

keep_ont_levels

Only keep phenotypes at certain absolute ontology levels to keep. See add_ont_lvl for details.

keep_onsets

The age of onset associated with each HPO ID to keep. If >1 age of onset is associated with the term, only the earliest onset is considered. See add_onset for details.

keep_tiers

Tiers from hpo_tiers to keep. Include NA if you wish to retain phenotypes that do not have any Tier assignment.

severity_threshold

Only keep phenotypes with severity scores below the set threshold. The severity score ranges from 1-4 where 1 is the MOST severe. Include NA if you wish to retain phenotypes that do not have any severity score.

pheno_frequency_threshold

Only keep phenotypes with frequency above the set threshold. Frequency ranges from 0-100 where 100 is a phenotype that occurs 100% of the time in all associated diseases. Include NA if you wish to retain phenotypes that do not have any frequency data.

keep_celltypes

Cell type to keep.

keep_seqnames

Chromosomes to keep.

gene_size

Min/max gene size (important for therapeutics design).

gene_frequency_threshold

Only keep genes with frequency above the set threshold. Frequency ranges from 0-100 where 100 is a gene that occurs 100% of the time in a given phenotype. Include NA if you wish to retain genes that do not have any frequency data.

keep_biotypes

Which gene biotypes to keep. (e.g. "protein_coding", "processed_transcript", "snRNA", "lincRNA", "snoRNA", "IG_C_gene")

keep_specificity_quantiles

Which cell type specificity quantiles to keep (max quantile is 40).

keep_mean_exp_quantiles

Which cell type mean expression quantiles to keep (max quantile is 40).

sort_cols

How to sort the rows using setorderv. names(sort_cols) will be supplied to the cols= argument and values will be supplied to the order= argument.

top_n

Top N genes to keep when grouping by group_vars.

group_vars

Columns to group by when selecting top_n genes.

phenotype_to_genes

Output of load_phenotype_to_genes mapping phenotypes to gene annotations.

hpo

Human Phenotype Ontology object, loaded from ontologyIndex.

return_report

If TRUE, will return a named list containing a report that shows the number of phenotypes/celltypes/genes remaining after each filtering step.

verbose

Print messages.

Value

A data.table of the prioritised phenotype- and celltype-specific gene targets.

Examples

res <- prioritise_targets()
#> Prioritising gene targets.
#> Importing existing file: ... phenotype_to_genes.txt
#> Adding term definitions.
#> Adding level-3 ancestor to each HPO ID.
#> Prioritised targets: step='start' 
#>  - rows: 419,111 
#>  - phenotypes: 5,443 
#>  - celltypes: 77
#> Filtering @ q-value <= 0.05
#> Prioritised targets: step='q_threshold' 
#>  - rows: 7,315 
#>  - phenotypes: 2,477 
#>  - celltypes: 77
#> Filtering @ fold-change >= 1
#> Prioritised targets: step='fold_threshold' 
#>  - rows: 7,315 
#>  - phenotypes: 2,477 
#>  - celltypes: 77
#> Getting absolute ontology level for 2,477 HPO IDs.
#> Prioritised targets: step='keep_ont_levels' 
#>  - rows: 7,315 
#>  - phenotypes: 2,477 
#>  - celltypes: 77
#> Annotating phenos with Onset.
#> Importing existing file: ... phenotype.hpoa
#> Importing existing file: ... phenotype.hpoa
#> Prioritised targets: step='keep_onsets' 
#>  - rows: 7,158 
#>  - phenotypes: 2,418 
#>  - celltypes: 77
#> Annotating phenos with Tiers.
#> Prioritised targets: step='keep_tiers' 
#>  - rows: 689 
#>  - phenotypes: 124 
#>  - celltypes: 66
#> Annotating phenos with Modifiers
#> Prioritised targets: step='severity_threshold' 
#>  - rows: 644 
#>  - phenotypes: 119 
#>  - diseases: 16 
#>  - celltypes: 66
#> Annotating phenotype frequencies.
#> Prioritised targets: step='pheno_frequency_threshold' 
#>  - rows: 558 
#>  - phenotypes: 99 
#>  - diseases: 15 
#>  - celltypes: 64
#> 33 / 64 of cell types kept.
#> Prioritised targets: step='keep_celltypes' 
#>  - rows: 352 
#>  - phenotypes: 86 
#>  - diseases: 15 
#>  - celltypes: 33
#> Filtering by gene size.
#> Converting phenos to GRanges.
#> Importing existing file: ... phenotype_to_genes.txt
#> Loading required namespace: ensembldb
#> Gathering gene metadata
#> Loading required namespace: EnsDb.Hsapiens.v75
#> Prioritised targets: step='keep_seqnames' 
#>  - rows: 57,639 
#>  - phenotypes: 86 
#>  - genes: 3,913
#> 234 / 3,913 genes kept.
#> Prioritised targets: step='gene_size' 
#>  - rows: 2,988 
#>  - phenotypes: 84 
#>  - genes: 234
#> Prioritised targets: step='keep_biotypes' 
#>  - rows: 2,988 
#>  - phenotypes: 84 
#>  - genes: 234
#> Filtering by specificity_quantile.
#> Filtering by mean_exp_quantile.
#> Annotating gene frequencies.
#> Importing existing file: ... genes_to_phenotype.txt
#> Prioritised targets: step='gene_frequency_threshold' 
#>  - rows: 12,054 
#>  - phenotypes: 83 
#>  - diseases: 14 
#>  - celltypes: 32 
#>  - genes: 233
#> Prioritised targets: step='keep_specificity_quantiles' 
#>  - rows: 312 
#>  - phenotypes: 59 
#>  - diseases: 12 
#>  - celltypes: 27 
#>  - genes: 67
#> Prioritised targets: step='keep_mean_exp_quantiles' 
#>  - rows: 312 
#>  - phenotypes: 59 
#>  - diseases: 12 
#>  - celltypes: 27 
#>  - genes: 67
#> Sorting rows.
#> Finding top 20 gene targets per: HPO_ID, CellType
#> Prioritised targets: step='top_n' 
#>  - rows: 312 
#>  - phenotypes: 59 
#>  - diseases: 12 
#>  - celltypes: 27 
#>  - genes: 67
#> Prioritised targets: step='end' 
#>  - rows: 312 
#>  - phenotypes: 59 
#>  - diseases: 12 
#>  - celltypes: 27 
#>  - genes: 67